requirements.txt
aactivator
black
django<=4.1
ipython
moviepy
opencv-python
pillow==9.2.0
rembg
requests==2.28.1
tqdm
validators